+++ public/css/style.css
@@ -1,0 +1,173 @@
+* {
+    box-sizing: border-box;
+    margin: 0;
+    padding: 0;
+    color: white;
+}
+
+header, main {
+    display: block;
+}
+
+header {
+    text-align: center;
+}
+
+.container {
+    display: flex;
+    flex-direction: row;
+}
+
+.columns {
+    float: left;
+    width: 250px;
+    padding: 8px;
+    text-align: center;
+}
+
+.container :nth-child(3) {
+    order: 1;
+}
+
+.container :nth-child(2) {
+    order: 3;
+}
+
+.container :nth-child(1) {
+    order: 2;
+}
+
+.price {
+    list-style-type: none;
+    border: 1px solid #eee;
+    margin: 0;
+    padding: 0;
+}
+
+.header, .buttonsDiv {
+    height: 70px;
+}
+
+.grey {
+    height: 120px;
+}
+
+.header {
+    background-color: #111;
+    font-size: 25px;
+}
+
+.price li {
+    border-bottom: 1px solid #eee;
+    padding: 20px;
+    text-align: center;
+}
+
+.buttonsDiv {
+    display: flex;
+    justify-content: space-evenly;
+}
+
+.buttonsDivSelected {
+    display: grid;
+    justify-content: space-evenly;
+    align-content: space-around;
+}
+
+.selected {
+    color: grey;
+    padding: 10px 20px;
+    background: rgba(0, 0, 0, 0);
+    border: rgba(0, 0, 0, 0);
+    font-size: 16px;
+    text-decoration: none;
+    text-transform: uppercase;
+    transition: .5s;
+    letter-spacing: 4px;
+    border-radius: 5px;
+    pointer-events: none;
+}
+
+html {
+    height: 100%;
+}
+
+body {
+    margin: 0;
+    padding: 0;
+    font-family: sans-serif;
+    background: linear-gradient(#141e30, #243b55);
+}
+
+.login-box {
+    position: absolute;
+    top: 50%;
+    left: 50%;
+    min-width: 790px;
+    max-height: 700px;
+    padding: 40px;
+    transform: translate(-50%, -50%);
+    background: rgba(0, 0, 0, .5);
+    box-sizing: border-box;
+    box-shadow: 0 15px 25px rgba(0, 0, 0, .6);
+    border-radius: 10px;
+}
+
+
+h1 {
+    margin: 0 0 30px;
+    padding: 0;
+    color: #fff;
+    text-align: center;
+}
+
+.login-box input {
+    outline: none;
+    border-bottom: none;
+}
+
+.login-box label {
+    top: 0;
+    left: 0;
+    color: #fff;
+    pointer-events: none;
+    transition: .5s;
+}
+
+.login-box label {
+    position: absolute;
+    padding: 10px 0;
+    pointer-events: none;
+    transition: .5s;
+    top: -20px;
+    left: 0;
+    color: #03e9f4;
+    font-size: 12px;
+}
+
+.button {
+    padding: 10px 20px;
+    color: #03e9f4;
+    background: rgba(0, 0, 0, 0);
+    border: rgba(0, 0, 0, 0);
+    font-size: 16px;
+    text-decoration: none;
+    text-transform: uppercase;
+    transition: .5s;
+    letter-spacing: 4px;
+    border-radius: 5px;
+}
+
+.button:hover {
+    background: #03e9f4;
+    color: #fff;
+    border-radius: 5px;
+    box-shadow: 0 0 5px #03e9f4,
+    0 0 15px #03e9f4,
+    0 0 15px #03e9f4,
+    0 0 15px #03e9f4;
+}
+
+.selected:hover {
+    /*cursor: default !important;*/
+}
