+++ public/loginAndForm/message.html
@@ -1,0 +1,22 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <link href="../../favicon.ico" rel="icon" type="image/x-icon">
+    <link href="../css/style.css" rel="stylesheet">
+    <script src="../js/message.js" type="module"></script>
+    <meta charset="UTF-8">
+    <title>feedback message</title>
+    <script
+            crossorigin="anonymous"
+            integrity="sha256-3zlB5s2uwoUzrXK3BT7AX3FyvojsraNFxCc2vC/7pNI="
+            src="https://code.jquery.com/jquery-3.6.1.js"></script>
+    <link crossorigin="anonymous" href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.0-beta3/dist/css/bootstrap.min.css"
+          integrity="sha384-eOJMYsd53ii+scO/bJGFsiCZc+5NDVN2yr8+0RDqr0Ql0h+rP48ckxlpbzKgwra6" rel="stylesheet">
+</head>
+<body>
+<div class="login-box">
+    <h2>Payment completed successfully</h2>
+</div>
+</body>
+</html>
+
